# wallpaper-server/API_DOCUMENTATION.yaml
openapi: 3.0.0
info:
  title: 壁纸系统API
  description: |
    壁纸系统后端接口文档
    
    ## 认证说明
    
    ### JWT Token认证
    系统使用JWT Token进行身份认证。登录成功后会返回Token，后续请求需要在请求头中携带Token。
    
    **请求头设置：**
    ```
    Authorization: Bearer <token>
    ```
    
    **示例：**
    ```
    Authorization: Bearer eyJhbGciOiJIUzI1NiJ9.eyJ1c2VySWQiOjEsInVzZXJuYW1lIjoiYWRtaW4iLCJyb2xlIjoiYWRtaW4iLCJpYXQiOjE3MDAwMDAwMDAsImV4cCI6MTcwMDA4NjQwMH0.xxxxx
    ```
    
    **Token有效期：**
    - 默认有效期：24小时（86400000ms）
    - 本地开发环境：7天（604800000ms）
    
    **Token过期处理：**
    - Token过期后，接口会返回 `401` 错误码
    - 错误信息：`登录已过期，请重新登录`
    - 需要重新调用登录接口获取新Token
    
    ### 兼容性说明
    为了兼容旧版本，系统同时支持 `X-User-Id` 请求头（仅开发环境使用）。
    
    ## 错误码说明
    
    ### 成功 (200)
    - 200: 操作成功
    
    ### 用户相关错误 (1000-1099)
    - 1001: 用户不存在
    - 1002: 用户已存在
    - 1003: 用户名或密码错误
    - 1004: 微信未绑定
    - 1005: 用户无权限
    - 1006: 用户暂无创作权限
    
    ### 壁纸相关错误 (2000-2099)
    - 2001: 壁纸不存在
    - 2002: 壁纸上传失败
    - 2003: 壁纸删除失败
    - 2004: 壁纸状态无效
    
    ### 创作者相关错误 (3000-3099)
    - 3001: 创作者不存在
    
    ### 分类相关错误 (4000-4099)
    - 4001: 分类不存在
    - 4002: 分类已存在
    
    ### 专辑相关错误 (5000-5099)
    - 5001: 专辑不存在
    - 5002: 专辑创建失败
    - 5003: 专辑删除失败
    
    ### 作品集相关错误 (8000-8099)
    - 8001: 作品集不存在
    - 8002: 作品集创建失败
    - 8003: 无权操作此作品集
    
    ### 用户行为相关错误 (6000-6099)
    - 6001: 已操作过
    - 6002: 操作记录不存在
    
    ### 举报相关错误 (7000-7099)
    - 7001: 已举报过
    - 7002: 举报失败
    
    ### 参数相关错误 (8000-8099)
    - 8001: 参数错误
    - 8002: 缺少必要参数
...
servers:
...
paths:
...
  /api/wallpaper/list:
    get:
      tags:
        - 壁纸管理
      summary: 获取壁纸列表
      description: 获取壁纸列表，支持按分类、类型筛选和分页，默认按创建时间降序排序。只返回可见的壁纸（已审核通过）
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
          description: 页码（从1开始）
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            default: 10
          description: 每页数量
        - name: type
          in: query
          required: false
          schema:
            type: integer
            enum: [1, 2, 3]
          description: 壁纸类型（1手机壁纸、2平板壁纸、3头像）
        - name: categoryId
          in: query
          required: false
          schema:
            type: integer
          description: 分类ID
        - name: userId
          in: query
          required: false
          schema:
            type: integer
          description: 用户ID
        - name: keyword
          in: query
          required: false
          schema:
            type: string
          description: 关键词搜索
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      list:
                        type: array
                        items:
                          $ref: '#/components/schemas/Wallpaper'
                      total:
                        type: integer
                        example: 100
                      page:
                        type: integer
                        example: 1
                      pageSize:
                        type: integer
                        example: 10
  /api/wallpaper/detail/{id}:
    get:
      tags:
        - 壁纸管理
      summary: 获取壁纸详情
      description: 根据壁纸ID获取壁纸详细信息
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 壁纸ID
        - name: userId
          in: query
          required: false
          schema:
            type: integer
          description: 当前用户ID（用于判断是否喜欢/收藏）
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/WallpaperDetail'
  /api/wallpaper/upload:
    post:
      tags:
        - 壁纸管理
      summary: 上传壁纸
      description: |
        创作者上传壁纸
        
        **需要认证**：请在请求头中携带 `Authorization: Bearer <token>`
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - url
                - type
                - title
              properties:
                url:
                  type: string
                  description: 原始图片URL（高清原图）【必填】
                thumbnailUrl:
                  type: string
                  description: 缩略图URL（列表展示用）
                imageWidth:
                  type: integer
                  description: 图片宽度（像素）
                imageHeight:
                  type: integer
                  description: 图片高度（像素）
                fileFormat:
                  type: string
                  description: 图片格式（JPG、PNG等）
                fileSize:
                  type: integer
                  description: 文件大小（字节）
                type:
                  type: integer
                  description: 壁纸类型（1手机壁纸、2平板壁纸、3头像）【必填】
                  enum: [1, 2, 3]
                title:
                  type: string
                  description: 壁纸标题【必填】
                categoryId:
                  type: integer
                  description: 分类ID
                description:
                  type: string
                  description: 壁纸描述
      responses:
        '200':
          description: 上传成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        '401':
          description: 未认证或Token过期
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: 未登录
                  data:
                    type: object
  /api/wallpaper/download/{id}:
    post:
      tags:
        - 壁纸管理
      summary: 下载壁纸
      description: |
        下载壁纸并记录下载行为
        
        **需要认证**：请在请求头中携带 `Authorization: Bearer <token>`
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 壁纸ID【必填】
      responses:
        '200':
          description: 下载成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        '401':
          description: 未认证或Token过期
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: 未登录
                  data:
                    type: object
  /api/wallpaper/like/{id}:
    post:
      tags:
        - 壁纸管理
      summary: 点赞壁纸
      description: |
        点赞或取消点赞壁纸
        
        **需要认证**：请在请求头中携带 `Authorization: Bearer <token>`
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 壁纸ID【必填】
      responses:
        '200':
          description: 操作成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        '401':
          description: 未认证或Token过期
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: 未登录
                  data:
                    type: object
  /api/wallpaper/collect/{id}:
    post:
      tags:
        - 壁纸管理
      summary: 收藏壁纸
      description: |
        收藏或取消收藏壁纸
        
        **需要认证**：请在请求头中携带 `Authorization: Bearer <token>`
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 壁纸ID【必填】
      responses:
        '200':
          description: 操作成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        '401':
          description: 未认证或Token过期
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: 未登录
                  data:
                    type: object
  /api/wallpaper/report/{id}:
    post:
      tags:
        - 壁纸管理
      summary: 举报壁纸
      description: |
        举报壁纸
        
        **需要认证**：请在请求头中携带 `Authorization: Bearer <token>`
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 壁纸ID【必填】
        - name: reason
          in: query
          required: true
          schema:
            type: string
          description: 举报原因【必填】
      responses:
        '200':
          description: 举报成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        '401':
          description: 未认证或Token过期
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: 未登录
                  data:
                    type: object
  /api/upload:
    post:
      tags:
        - 文件上传
      summary: 上传图片
      description: 上传图片文件，支持手机壁纸、平板壁纸、头像三种类型
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: 图片文件【必填】
                type:
                  type: integer
                  description: 图片类型（默认1）
                  enum: [1, 2, 3]
                  default: 1
                  example: 1
      responses:
        '200':
          description: 上传成功
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    description: 手机壁纸/平板壁纸返回
                    properties:
                      url:
                        type: string
                        description: 原始图片URL（高清原图）
                        example: https://wallpaper-project.oss-cn-beijing.aliyuncs.com/phone/abc123.jpg
                      thumbnailUrl:
                        type: string
                        description: 缩略图URL（列表展示用）
                        example: https://wallpaper-project.oss-cn-beijing.aliyuncs.com/phone/abc123.jpg?x-oss-process=image/resize,w_300,h_300
                      width:
                        type: integer
                        description: 图片宽度（像素）
                        example: 1920
                      height:
                        type: integer
                        description: 图片高度（像素）
                        example: 1080
                      format:
                        type: string
                        description: 图片格式（JPG、PNG等）
                        example: JPG
                      size:
                        type: integer
                        description: 文件大小（字节）
                        example: 2048000
                  - type: object
                    description: 头像返回
                    properties:
                      url:
                        type: string
                        description: 头像URL
                        example: https://wallpaper-project.oss-cn-beijing.aliyuncs.com/avatars/abc123.jpg
...
  /api/wallpaper/upload-to-work:
    post:
      tags:
        - 壁纸管理
      summary: 上传壁纸到作品集
      description: 将壁纸上传到指定的作品集，需要登录且是作品集作者
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                workId:
                  type: integer
                  description: 作品集ID
                  example: 1
                url:
                  type: string
                  description: 图片URL
                  example: /static/images/wallpaper1.jpg
                thumbnailUrl:
                  type: string
                  description: 缩略图URL
                  example: /static/images/thumbnails/wallpaper1.jpg
                imageWidth:
                  type: integer
                  description: 图片宽度（像素）
                  example: 1080
                imageHeight:
                  type: integer
                  description: 图片高度（像素）
                  example: 1920
                fileFormat:
                  type: string
                  description: 文件格式
                  example: JPG
                fileSize:
                  type: integer
                  description: 文件大小（字节）
                  example: 2048000
      responses:
        '200':
          description: 上传成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: success
                  data:
                    type: integer
                    description: 壁纸ID
                    example: 123
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        '8001':
          description: 作品集不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        '8003':
          description: 无权上传到此作品集
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Token认证，格式为 `Bearer <token>`
  schemas:
    Result:
      type: object
      properties:
        code:
          type: integer
          example: 200
          description: 状态码（200成功，其他为错误码）
        message:
          type: string
          example: success
          description: 返回消息
        data:
          type: object
          nullable: true
          description: 返回数据
...
    Wallpaper:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: 动漫壁纸
        url:
          type: string
          description: 原始图片URL（高清原图）
          example: /static/images/wallpaper1.jpg
        thumbnailUrl:
          type: string
          description: 缩略图URL（列表展示用）
          example: /static/images/wallpapers/thumbnails/wallpaper1.jpg
        imageWidth:
          type: integer
          description: 图片宽度（像素）
          example: 1080
        imageHeight:
          type: integer
          description: 图片高度（像素）
          example: 1920
        fileFormat:
          type: string
          description: 图片格式（JPG、PNG等）
          example: jpg
        fileSize:
          type: integer
          description: 文件大小（字节）
          example: 2048000
        type:
          type: integer
          example: 1
        category:
          type: string
          description: 分类名称
          example: 动漫次元
        description:
          type: string
          example: 这是一张动漫壁纸
        author:
          type: object
          properties:
            id:
              type: integer
              example: 1
            name:
              type: string
              example: 夏天
            avatarUrl:
              type: string
              example: /static/images/avatar1.jpg
        viewCount:
          type: integer
          example: 500
        downloadCount:
          type: integer
          example: 100
        likeCount:
          type: integer
          example: 100
        collectCount:
          type: integer
          example: 50
        isLiked:
          type: boolean
          example: true
        isCollected:
          type: boolean
          example: false
        createTime:
          type: string
          format: date-time
          example: 2026-03-16T10:00:00
    WallpaperDetail:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: 动漫壁纸
        url:
          type: string
          description: 原始图片URL（高清原图）
          example: /static/images/wallpaper1.jpg
        thumbnailUrl:
          type: string
          description: 缩略图URL（列表展示用）
          example: /static/images/wallpapers/thumbnails/wallpaper1.jpg
        imageWidth:
          type: integer
          description: 图片宽度（像素）
          example: 1080
        imageHeight:
          type: integer
          description: 图片高度（像素）
          example: 1920
        fileFormat:
          type: string
          description: 图片格式（JPG、PNG等）
          example: jpg
        fileSize:
          type: integer
          description: 文件大小（字节）
          example: 2048000
        type:
          type: integer
          example: 1
        description:
          type: string
          example: 这是一张动漫壁纸
        viewCount:
          type: integer
          example: 500
        downloadCount:
          type: integer
          example: 100
        likeCount:
          type: integer
          example: 100
        collectCount:
          type: integer
          example: 50
        isLiked:
          type: boolean
          example: true
        isCollected:
          type: boolean
          example: false
        author:
          type: object
          properties:
            id:
              type: integer
              example: 1
            name:
              type: string
              example: 夏天
            avatarUrl:
              type: string
              example: /static/images/avatar1.jpg
            bio:
              type: string
              example: 个人简介
            followerCount:
              type: integer
              example: 100
            workCount:
              type: integer
              example: 10
...